Highlight non-ASCII characters
I've been bitten occasionally in the past by Unicode characters that
appear similar to their ASCII equivalents, e.g. a space, which can
sometimes produce great confusion for program runtimes or compilers, as
well as me.

To make it obvious when a non-ASCII character is used, highlight it. I'm
using Vim's built-in 'ErrorMsg' highlight colour, which is red.

See also:
https://github.com/reinderien/mimic